third_party/WebKit/LayoutTests/svg/custom/createelement-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: Cannot read property 'setAttribute' of null
